Moving frequently used key to the thumb cluster The idea is to "unload" keys that were by default layout set to be used by the pinky to the thumb cluster by using the hold-tap functionality.
For example Ctrl keys are held while the ESC key is pressed so
the two can be combined into a single key.

For combination such as Ctrl/ESC I am using the default mod tap
which will trigger Ctrl If you press another key within the 200ms
of it being pressed. This is done to enable combinations such as
Ctrl+c/Ctrl+v where Ctrl is used as a modifier but it is not intended
to be held but inside used in a "rolling" motion.

Unlike the mod tap I am using for Shift/Enter which is the opposite
and preference the tap and shift won't gets triggered if another key
is pressed within 200ms of it being pressed.