Added capsword zmk functionality
I already have defined the ESC key to be part of the thumb cluster
in the past to shift some of the load from the pink pinky finger.

This means I was no longer using the factory set ESC key which
happened to be at the same place where CAPS key is located in
a traditional keyboard layout. So I now have a "CAPS" key
that is much more convenient then the traditional one.