Updated key keymap to move from V2.0 to V3.0
Note: version are stated as the implementation of zmk done
by kinesis for the advantage 360 pro and not the versions used
by zmk.

The big issue I came across is the fact that all my changes
to the adv360.keymap file were done using no longer supported
keymap which means that when I tried to rebase my commits unto
the V3.0 branch, not only did I get merge conflicts, but all of
my commits were conflicting with the base I was now trying to
use. The dirty "fix" was to override the new keymap with the commits
that were conflicting with it and add one final commit that fixes
the keymap to comply with the one in V3.0. Then I squashed all of
my commits into this one commit.

History of my commits prior to the move to V3.0 can be found
on my V2.0 branch.