Hold tap on the single quote key
I not only added mod-tap(my home raw mode flavor - hm) but I also
decided that the default behavior is now the double quote as this
is what I use more then the single quote.

When key is tapped a double quote is produced.
When key is held a single quote is produced.