Created a basic layer for tmux use This should make the use of tmux more ergonomic.
Some implementation details:
- I could have created a macro that executes the prefix for tmux and
  then switches to the tmux layer but this means that if the layer
  timed out and I did not used it I would be with tmux in "command mode"
  and with no tmux lyaer. This could have been solved by setting a
  timeout in the tmux settings(tmux.conf) for the prefix being active.
  and I still haven't crossed this option off, my main concern is
  coupling keyboard configuration to tmux.conf.
- The previous issue could have been solved if instead of executing the
  prefix macro and going into the tmux layer I would start by executing
  ESC and then go into the prefix and switch to the tmux lyaer(sticky
  layer). But this too has its limitations, for example vim would go
  into normal mode each time I would execute a tmux command, so again
  more coupling.

In the end I went with backing the prefix into the layer itself. This
creates issue in the form of creating multiple macros. ZMK does
support the use of parameterized macros but I was not able to make it
work with the Adv360-Pro-ZMK fork. Once I am able to do so, I will be
able to limit the number of macros I have defined and strike a good
balance between extra code vs coupling my settings to depend on external
software.